% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfnswapi-package.R
\docType{package}
\name{tfnswapi-package}
\alias{tfnswapi}
\alias{tfnswapi-package}
\title{tfnswapi: Access TfNSW Open Data from R}
\description{
Access TfNSW Open Data from R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/asiripanich/tfnswapi}
  \item Report bugs at \url{https://github.com/asiripanich/tfnswapi/issues}
}

}
\keyword{internal}
