% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tfnswapi_get_api_key}
\alias{tfnswapi_get_api_key}
\title{Get registered API key.}
\usage{
tfnswapi_get_api_key()
}
\value{
Character.
}
\description{
Get registered API key.
}
\examples{

tfnswapi_get_api_key()
}
