% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{tfnswapi_register}
\alias{tfnswapi_register}
\title{Register TfNSW API key.}
\usage{
tfnswapi_register(key, add_to_renviron = FALSE, overwrite = FALSE)
}
\arguments{
\item{key}{Character. TfNSW API key.}

\item{add_to_renviron}{Logical. Default is \code{FALSE}. If \code{TRUE} then the file directory
of the authentication file will be added to your \code{.Renviron} file.}

\item{overwrite}{Logical. Default is \code{FALSE}. If \code{TRUE}, the existing \code{TFNSW_API_KEY}
R environment variable will be replaced with new \code{key}.}
}
\description{
Register TfNSW API key.
}
\examples{
\dontrun{

tfnswapi_register(key = "your-key")

}
}
