% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{tfnswapi_get}
\alias{tfnswapi_get}
\title{Get data using GET API.}
\usage{
tfnswapi_get(api, params = NULL, key = tfnswapi_get_api_key())
}
\arguments{
\item{api}{Character. API name (e.g. 'carpark' for
https://opendata.transport.nsw.gov.au/node/6223/exploreapi#!/carpark/GetCarPark).}

\item{params}{a named list. Any parameters that the API accepts can be specified
here.}

\item{key}{Character (optional). An API key can be provided, else if emptied
it will look for any previously registered API key.}
}
\value{
a response list
}
\description{
Get data using GET API.
}
\examples{

\dontrun{

 # See what facilities are available
 tfnswapi_get("carpark")
 tfnswapi_get("carpark", params = list(facility = 2))

}
}
